# openapi.yaml
openapi: 3.0.0
info:
  title: ORIGINE.DIGITAL API
  version: 1.0.0
  description: Autonomous digital shop API for clients and third‑party integrators.
servers:
  - url: https://your-domain.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
security:
  - bearerAuth: []
paths:
...
  /content/social:
    get:
      summary: List social posts
    post:
      summary: Generate a social post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                topic: { type: string }
                platform: { type: string, enum: [twitter, linkedin, facebook, instagram] }
                tone: { type: string }
...
security:
  - bearerAuth: []
...